pkg/destroy/aws: Pull ARN parsing up out of deleteARN()
This gives us structured information in the caller, which I'll take
advantage of in a follow-up commit.  Also pull the logger generation
up into the caller, to follow our "functions do not WithField their
arguments, since the caller could do that" pattern.  This commit is a
pure refactor with no functional change.