pkg/asset/installconfig/platform: Link to platform list
For folks wondering "hey, why isn't the wizard asking me about
OpenStack?", etc.  Experimental and user-provided platforms are
unreliable and/or complicated enough that we are keeping them out of
the wizard to avoid confusing folks using the wizard for our reliable,
supported platforms.

The repo link is a somewhat sneaky way to get folks in front of the
platform list that leads the README [1].  Alternatives considered [2]
included:

  https://github.com/openshift/installer/blob/master/README.md#supported-platforms

but that floating link could diverge from the actual installer that
the user was using (e.g. if a new platform was added in master, it
would show up in the linked README, but not be supported by the users
old installer).  We also considered:

  https://github.com/openshift/installer/blob/{commit}/README.md#supported-platforms

but were concerned about commit divergence due to build tooling.  [3]
should address our current divergence, but that sort of thing is
probably too brittle for docs where we need high reliability.  The
long-term plan is probably linking out to some more official docs
outside of the installer repository, but this repo link should help
until then.

[1]: https://github.com/openshift/installer/blob/c6517384e71e5f09931c4da5e772fdec225d02ec/README.md#supported-platforms
[2]: https://github.com/openshift/installer/pull/1713#discussion_r286109922
[3]: https://github.com/openshift/installer/pull/1744