Generate config for OpenStack cloud provider
This commit allows the provider to get a configuration from the secret.

The in-tree cloud provider can read the configuration from either a file
on the local file system or a secret. In order to read the data from the
secret, three parameters must be written into the local configuration:
the name of the secret, the name of the namespace containing the secret,
and the path to the kubeconfig file.

Example:
$ cat /etc/kubernetes/cloud.conf
[Global]
secret-name = openstack-credentials
secret-namespace = kube-system
kubeconfig-path = /var/lib/kubelet/kubeconfig

Based on these three parameters, the in-tree provider reads the real
configuration with cloud credentials from the secret under the key
"clouds.conf".

Enabling the provider is done by adding --cloud-provider=openstack
and --cloud-config=/etc/kubernetes/cloud.conf to the kubelet command.