Generate an `ignition-auth` key and provide it to the MCS
This is an optional hardening for access to Ignition; the installer
generates a random key (separately for master/worker pool) and installs
it into the `openshift-machine-config-operator` namespace.  If the MCS
finds an `ignition-auth` secret with the `master/worker` keys, it will use it:
https://github.com/openshift/machine-config-operator/pull/736

This PR just generates those secrets, so we can land it before the
MCO PR as well.