terraform: change override directory for init to `.tf`
With terraform 0.12 `init` behavior has changed to re-use the state file when initializing. With the previous override directory set to the working directory of terraform, init would assume initializing has already occured and fail to load the state file, when it should not try to load the state file.

Moving the override directory to `<working dir>/.tf` forces terraform to correctly initialize