terraform: bump the vendored version to 0.12-rc.1
Terraform community is getting ready for 0.12 release that introduces wide variety of improvements in the language
and fixes a lot of the subtle plan and apply bugs. see here [1] for official announcement

New platforms like Azure require that installer moves to 0.12 version of terraform as it provides the required `dynamic` blocks in HCL2 [2]. For example,
Azure `dns_srv` record requires use of `foreach` to create dynamic count of records for control plane [3]. `RHHI` platform owners have also requested 0.12 terraform version to
simplify their terraform templates. Therefore, moving the installer to 0.12-rc.1 sets us up to switch to final release of 0.12 which is very close [4].

This commit,
- moves the vendored copy to v0.12.0-rc1 [5] tag revision.
- expands the list of required packages from terraform helper packages as these are required by newer terraform-providers
- dep is sadly unsuccesful in understanding the correct version for the indirect dependencies, and therefore it was required to add overrides for faulty vendored packages based on upstream go.mod [6] ang go.sum [7]

[1]: https://www.hashicorp.com/resources/introducing-terraform-0-12
[2]: https://www.terraform.io/upgrade-guides/0-12.html#attributes-vs-blocks
[3]: https://github.com/openshift/installer/pull/1454/files#diff-764adb23dcc0edbbebc09192eb233e9aR46
[4]: https://github.com/hashicorp/terraform/pulls?q=is%3Aopen+is%3Apr+milestone%3Av0.12.0
[5]: https://github.com/hashicorp/terraform/releases/tag/v0.12.0-rc1
[6]: https://github.com/hashicorp/terraform/blob/v0.12.0-rc1/go.mod
[7]: https://github.com/hashicorp/terraform/blob/v0.12.0-rc1/go.sum