WIP: Bug 1711844: bootkube.sh: Use pause image from payload
A while ago we switched to pulling the pause image from the payload,
this fixes things to do the same for the bootstrap node.  We need
this to support disconnected installs, countries where gcr.io is firewalled,
and further we really only want to consume our own release payload, and
not trust external registries.