data/azure: use the credentials provided to installer for creating resources
Terraform azurerm provider supports supports 4 kinds of authentication methods [1].
The installer uses the `Service Principal and a Client Secret` [2] method because it best suites the requirement for installer
as this type of suth methods can be passed into the cluster as secret.

By default when no authentication is specified for Azurerm provider, the ordering followed is (service principal + client secret) > (service principal + client secret) > (msi) > (azure cli token) [3]

Currently since the provider is not setup with any authentication in the templates, it usually falls back to the azure cli method.

Now since installer sets up the creds for users, there is no way to use the ENV input for (service principal + client secret) and terraform
Azurerm provider doesn't support file based source for creds.

Therefore, they need to be explicitly provided by the installer to the templates.

[1]: https://www.terraform.io/docs/providers/azurerm/#authenticating-to-azure
[2]: https://www.terraform.io/docs/providers/azurerm/auth/service_principal_client_secret.html
[3]: https://github.com/hashicorp/go-azure-helpers/blob/5e51ac013932f1c9434224349e74d77be738739c/authentication/builder.go#L49-L56