data/azure: create an explicit dependency on private zone before VMs are created
From the Private DNS Overview Doc [1]

```
The virtual network needs to be empty (that is, no VM records exist) when it initially (that is, for the first time) links to a private zone as a registration or resolution virtual network. However, the virtual network can then be non-empty for future linking as a registration or resolution virtual network, to other private zones.
```

Therefore, it looks like there needs to be an explicit dependency between private zone and VMs being created in the VNET.

Moving the private zone resource to main.tf and making bootstrap, masters, dns module consume a variable with value from private zone resource, makes terraform create the DNS Zone before creating any resources from those modules.

There is hope that the constraint might be lifted when private DNS zones become publicly GA. [2]

[1]: https://docs.microsoft.com/en-us/azure/dns/private-dns-overview
[2]: https://feedback.azure.com/forums/217313-networking/suggestions/35340511-create-private-dns-zone-in-virtual-network-which-a