azure: allow installer to consume `AZURE_AUTH_LOCATION` env for credentials file
Installer uses `NewAuthorizerFromFileWithResource` [1], which uses `GetSettingsFromFile` [2] to locate and load the file with auth credentials.
`GetSettingsFromFile` [2] uses the `AZURE_AUTH_LOCATION` env [3] to locate the file with no way to override or specify explicitly.

Currently the installer uses the hard-coded location `~/.azure/osServicePrincipal.json` to load the credentials. But for CI, it would be important to override this
location to another location like we do for AWS [4]. So this change allows users to set `AZURE_AUTH_LOCATION` env to provider installer custom location to auth file.

[1]: https://github.com/Azure/go-autorest/blob/v12.0.0/autorest/azure/auth/auth.go#L243
[2]: https://github.com/Azure/go-autorest/blob/v12.0.0/autorest/azure/auth/auth.go#L287
[3]: https://github.com/Azure/go-autorest/blob/v12.0.0/autorest/azure/auth/auth.go#L289
[4]: https://github.com/openshift/release/blob/6c0b409639d6dcd074238e5396cddcc5c4da1510/ci-operator/templates/openshift/installer/cluster-launch-installer-e2e.yaml#L373-L374