Text edits for the CloudFormation templates
Because we can no longer use "UPI" in customer-facing documentation and the CloudFormation templates in /installer are published in /openshift-docs, I want to change the source templates instead of having to always reproduce the fix in the docs repo.

I'm making a few other style and changes to the parameter descriptions.

Per a comment on a docs PR (https://github.com/openshift/openshift-docs/pull/15059), the subnet type for the worker CloudFormation template  needs to change to PrivateSubnet from WorkerSubnet. Trevor clarified that the subnet in the worker does not need to be private. s/PrivateSubnet/Subnet in just that CloudFormation Template.