data/azure: clarify that the password isn't used
The Azure ARM API requires that a username and password are provided
when creating a VM, even if it isn't going to be used. RHCOS doesn't
actually apply this password because WALA (the Azure Linux agent) isn't
shipped in the OS. This change just aims to make this fact more clear to
future readers.