pkg: add gather/ssh package for utilities for gathering using ssh
The package allows users to create a new SSH client that
- Uses an SSH agent to provide authentication
- Either uses the keys from paths explicity provided, or loads all the keys from `~/.ssh` to use for authentication

The package also provides a way to Run a command using an SSH client. The client must be configured to allow SSH agent forwarding.

The package also provides a way to pull a file from remote server to a file on the local host using a SSH client.