cmd: gather the logs from bootstrap instead of printing commands
This changes the gather bootstrap to establish an SSH connection to bootstrap host to gather the logs, rather than printing the commands for
the user to execute.

This had some trivial conflicts because e4bddc214e (gather: add
support for azure, 2019-05-28, #1795) had landed in release-4.2 but
does not apply to 4.1, where we have no Azure support.