hack/build: Use BUILD_VERSION if non-empty
Like 3313c08266 (hack/build: Use SOURCE_GIT_COMMIT if set, 2019-05-13, #1828).
This should fix [1]:

  $ ./openshift-install version
  ./openshift-install v4.1.0-201905212232-dirty
  ...

which is from Git looking at the build directory (and seeing -dirty
because doozer is adjusting our Dockerfile?).  Since the tags aren't
in the source repository (github.com/openshift/install), we need to
use the BUILD_* variables which were added to Doozer in [2].  In a
recent build [3], these looked like:

  ENV SOURCE_GIT_COMMIT=8aa5b10aa82d201deba0befbfac9fabf76f719f4 SOURCE_DATE_EPOCH=1561485623 BUILD_VERSION=v4.1.9 SOURCE_GIT_URL=https://github.com/openshift/installer SOURCE_GIT_TAG=8aa5b10a BUILD_RELEASE=201907311355

With this commit, we prefer BUILD_VERSION, falling back to 'git
describe' in the local repository if BUILD_VERSION is unset or empty.

Doozer also sets BUILD_RELEASE, but Luke says we always set
BUILD_VERSION in OpenShift 4 so there's no need to include it in the
fallback chain [4] (otherwise I'd have preferred BUILD_VERSION,
falling back to BUILD_RELEASE, falling back to Git inspection).

I don't see a need to use:

  GIT_TAG="${BUILD_VERSION}-${BUILD_RELEASE}"

with both, because if we know the version is v4.1.8, who cares about
the timestamp?  We only ever get a single 4.1.8 far enough along to
show up in front of end-users, even if there's a build hiccup or some
such that causes us to build multiple rounds with the same version
internally.  There's a bunch of v4.2.0, etc., in the run-up to a new
minor release, but we have commit hashes to distinguish between those.

[1]: https://github.com/openshift/installer/issues/1828
[2]: https://gitlab.cee.redhat.com/openshift-art/tools/doozer/commit/b02114a0bb26bbb7f60f1cc484c17fc3172c6df4
     distgit.py: ART-165 add build ENV vars, 2019-07-07
[3]: http://pkgs.devel.redhat.com/cgit/containers/ose-installer/tree/Dockerfile?h=rhaos-4.1-rhel-7
[4]: https://github.com/openshift/installer/pull/1829#discussion_r310290990