data/data/openstack: Open UDP 5353 (mdns) port
The mdns port is going to be used for updating the internal DNS
records necessary for the installation to succeed (the master & etcd A
as well as the SRV records).

Some OpenStack deployments allow UDP traffic by default, but some
don't so we need to create this Security Group Rule explicitly.