cmd/openshift-install/gather: Guard against cluster-side artifact collisions
Support running multiple gather commands without cluster-side
collisions.  For example:

1. Alice runs gather bootstrap, her command writes a tarball on the
   bootstrap machine and starts downloading it.
2. Bob runs gather bootstrap, his command starts writing to the same
   tarball on the bootstrap machine.
3. Alice's download croaks on the truncated file.

But maybe Alice's download already has an open file descriptor, so
Bob's clobber doesn't break the download?  I guess it depends on the
specifics of how installer-gather.sh's > ~/log-bundle.tar.gz worked.
With this commit, we avoid potential issues by passing in an optional
gather ID which the cluster-side scripts use to uniquify their
collection directories.  We still race if multiple gathers are
launched in the same second, but that's at least a narrower window,
and we can close it later by using a better source of random
characters.