asset/cluster: inject installer info when creating
OpenShift telemetry is currently unable to report _how_ a cluster was
installed (i.e. user-provisioned vs installer-provisioned). We currently
use a heuristic to try to determine this information.

This patch adds a ConfigMap immediately before cluster creation with
information about the installer. This ConfigMap is not present when the
user provisions the cluster themselves since it isn't injected until the
Terraform variables are generated. This ConfipMap contains the installer
version and the name of the invoker (can be overridden with the
OPENSHIFT_INSTALL_INVOKER environment variable). This will allow us to
determine whether a cluster was user-provisioned or installer-
provisioned, whether or not a custom build of the installer was used,
and who created the cluster (e.g. Hive, CI).

This commit cherry-picks 253126f597 (asset/cluster: inject installer
info when creating, 2019-06-21, #1890), although it's not a clean pick
because of conflicting context in the imports between the master
branch and this release-4.1 line [1].

[1]: https://github.com/openshift/installer/pull/1890#issuecomment-516537931