issue-template: Make it easy to remove irrelevant platforms
With the current template, people just fill in the name of their platform,
leaving the list of platforms as is. This makes it a bit hard to label
issues. Let's modify the template so the platforms are mentioned in a
comment and user explicitly ask writes (or copy&paste) the relevant
platform.