pkg/asset/installconfig: Allow users to provide the Zone ID for BaseDomain
Cloud providers usually allow users to create multiple DNS zones for the
same name, and users can configure their domain with registrars to point
to a specific DNS Zone's nameservers.

So if there are multiple DNS zones with BaseDomain as their name, the
installer could end up picking the wrong DNS Zone. Such errors can be
difficult to debug.

This fix will allow users to pick the ZoneID when there are multiple
DNS Zones with the same name.

This only addresses AWS and Azure at the moment and is a WIP. Tests
and validators need to be added.

https://jira.coreos.com/browse/CORS-1070