OpenStack: fix cluster destroy
The load balancer destroy step on the cluster destroy is expecting
for the cloud to have octavia service enabled, in case it is not
enabled the destroy command will fail. This commit fixes the issue
by skipping load balancers deletion in case no octavia endpoints
are enabled, which avoids a fatal error.

Fixes #1919