data/azure: use single storage account for cluster
Currently we use a storage account for boot diagnostics for the masters, another one for storing the ignition config for bootstrap node.
In the future we might need to store the RHCOS vhd in a storage account, and there is no valid case for these many storage accounts created by installer. And each storage container can have its invidual ACL.

Using a single storage account is simpler.