data/azure: create an image for booting machines for the cluster from vhd blob
previously the terraform expected an Azure Image [1] provided to the cluster, but that approach had some disadvantages:
* the image needed to be created before creating the cluster
* the image location was hardcoded to a specific resource-group

And since RHCOS isn't avaiable on the Public Registry of images, the new flow is using the vhd blob from Azure storage and creating an Azure Image from it for the cluster.

The public vhd is copied to cluster's resource group to make sure the image is not based directly on a public artifact.