pkg: update the machines and tfvars to correct image
Currently the OS image source for Azure was hard coded to `/resourceGroups/rhcos_images/.../rhcos` in rhcos.Image asset which means the image needed to pre-exist in the same sunscription.

But now the OS image source for Azure will be a vhd blob in Azure storage account as specified by build pipeline and the machines will be using the image in the cluster's resource group.