data/gcp/network: create nat-gateways per subnet for egress from private instances
Based on internet-access-reqs docs [1], the private instances need Cloud-NAT [1] resources to allow egress to internet.

The Cloud NAT gateways provide an automatically scaling-ha way to access the internet, so creating one per subnet (master and worker) should be
enough.
Cloud NATs also require a Cloud Router [3] for the control-plane, no data flows through Cloud router. This change creates one Cloud Router for the cluster.

Both the resources are regional and created in the same region as the cluster.

[1]: https://cloud.google.com/vpc/docs/vpc#internet_access_reqs
[2]: https://cloud.google.com/nat/docs/overview
[3]: https://cloud.google.com/router/docs/concepts/overview