[gcp] pkg/destroy: handle async call results, log delete info messages
Updates GCP destroy to handle async call results by using the same
request id every time a particular item is deleted. The resulting
operation should return the latest results of the particular request.
It updates the destroy function interface to return a boolean indicating
whether there are items still pending to be deleted. It also returns a
slice of errors instead of a single error so that all items of a
particular type are attempted instead of up to the first failure.
Info log statements are now logged when an item is successfully deleted.