scheduler: Use schedulable masters if no compute hosts defined.
This change makes use of a new configuration item on the scheduler CR
that specifies that control plane hosts should be able to run
workloads.  This option is off by default, but will now be turned on
if there are no compute machine pools with non-zero replicas defined.

This change also removes a validation and warning when no compute
hosts are defined, as an install with this configuration will now
complete successfully.