scheduler: Allow configuration of schedulable masters.
This change makes use of a new configuration item on the scheduler CR
that specifies that control plane hosts should be able to run
workloads.  This option is off by default, but can be changed through
a new install-config option under controlPlane.

The justification for a new install-config option for this is that
this is a required tunable for some clusters to have a successful
install.  A 3-node cluster on bare metal for example will need
schedulable masters.