data/data/openstack: Add registry to coredns image
The OpenStack deployment started failing with the following message:

Error: error pulling image "openshift/origin-coredns:v4.0": unable to pull openshift/origin-coredns:v4.0: image name provided is a short name and no search registries are defined in the registries config file.

That caused the `local-dns` service on the Service VM to fail which
resulted in none of the nodes receiving Ignition and booting up.

Specifying the `quay.io` registry explicitly fixes this issue.