bootkube: add static pod images to bookube flags
These images are used by openstack and KNI in order to set up some
infra services that are not provided in our case. This includes node
level DNS and loadbalancing, as well as keepalived to provide failover
for these services.

This requires the flags to be added in MCO first:
https://github.com/openshift/machine-config-operator/pull/983

The images are used in these MCO PRs:
https://github.com/openshift/machine-config-operator/pull/795
https://github.com/openshift/machine-config-operator/pull/740