baremetal: keepalived/coredns containers remove before create
If the container exists unexpectedly for any reason, then
ExecStop not called, and sometimes there are "storage for container removed"
errors trying to re-start it, presumably because podman cleanup
removes resources for exited containers.

So we add --rm on the podman create, and add logic to be completely sure any
stale containers are removed on startup of the services.

Closes: #2032