Deploy baremetal bootstrap services with MCO
Currently we have two very different sets of manifests and templates
for deploying the baremetal infrastructure services. One lives in the
installer, the other lives in MCO. This change removes the services
from the installer and makes use of the files being added to the MCO
that serve the same purpose but keep all of the templates in the same
repo and allow us to use the same MCO data for populating dynamic
values.

On platforms that don't need these services on the bootstrap node,
MCO will not render these files. We should check for the existence
of the relevant manifest and if present try to copy those files.

The static-pod-resources are grouped with the manifest because the
manifests require them, so it would be an error for one to exist
without the other.