baremetal: Decouple rhcos.Image for bootstrap/cluster
For the baremetal platform we require two different bootimages,
the QEMU one for the libvirt based bootstrap VM, and the OpenStack
one that contains the necessary Ironic config drive support to pass
data to ignition.

So we rework the Image abstraction by adding a new BootstrapImage
type/asset, which returns the same as Image in all cases except for
the baremetal platform.

This also aligns with the tfvars renamed in #2044 and allows us to
pass the rhcos.QEMU image via BootstrapImage to terraform but leaves
the OpenStack image URL available for future use to deploy masters
via follow-up PRs that implement issue #2060 and also correctly set
the worker machineset providerSpec for the baremetal-operator.

Related: #2037