gcp: reduce the length of the names of machines
The GCE kubelet uses the `os.Hostname()` to create node names and there's a har limit on hostnames to be 64 characters. based on [1]
On GCP, the hostnames are `<instance name>.c.<project-id>.internal` and since the project ids on GCP can be of max 30 characters, an instance name can be max 22 characters.
The instances created by a machineset have the `6` character of randomness added to machineset's name, therefore the max length of machineset's name can be 16 characters. Therefore to accomodate the pool's name and zone (as we create machineset per zone) :

the mastes machines are now `<infra-id>-m-<idx>` and
the worker machinesets are now `<infra-is>-w-<zone>`, where the infra-id can be max 12 characters on GCP.

[1]: https://tools.ietf.org/html/rfc2181#section-11