Conditionally include the bootstrap node in the SRV record
Define a new tfvar `etcd_pivot` that is set based on the
OPENSHIFT_INSTALL_ETCD_ENABLE_PIVOT environment variable, and controls
the addition of the bootstrap node to the etcd srv record.

This work is part of the redesign of how etcd is deployed to masters,
and the existance of the bootstrap node in this record will initially
be used by bootkube.sh to determine if this new functionally should be
enabled, and like on masters is used to validate that the bootstrap node
is permitted to host etcd.

The patch includes support for AWS and libvirt