OpenStack: allow to disable assigning of a floating ip to the bootstrap machine
Floating IP (fip) on the bootstrap machine is used to collect logs only, so,
it's optional, and OCP can operate without it.

In some cases, when the amount of fips is limited we can't set an ip and
the whole process fails. To prevent this this commit adds a boolean variable
"enable_bootstrap_floating_ip" that allows to disable assigning of a
floating ip to the bootstrap machine.