[gcp] Fix deletion of internal/external load balancers
External load balancers are not getting cleared because the function to
extract name and zone from an instance URL is returning name and zone in
reverse order. Clearing health checks should return immediately if no
health checks are found. Fixed the name of the node health check getting
deleted as part of the external load balancers.

For internal load balancers, it is not necessary to clear health checks
from backend services. Removed that code.