Add neutron ports for api and apps VIPs
Previously the service VM would proxy all traffic to the masters and
workers.

With this new architecture the worker nodes are addressed via an
ingress VIP local to the cluster, and need to be reachable from the
outside in order to serve the *.apps. We're attaching a floating IP to
the port, that maps to the ingress VIP shared among the workers.